# openapi.yaml
openapi: 3.0.3
info:
  title: AI 食物卡路里辨識系統 API
  description: |
    一套完整的 AI 識別食物卡路里 Web 應用程式 API
    包含會員系統、拍照辨識、飲食記錄、數據分析等功能
  version: 1.0.0
  contact:
    name: API Support
    email: support@calorie-ai.com
servers:
  - url: https://api.calorie-ai.com/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /users/me:
    get:
      tags:
        - User
      summary: 取得當前會員資料
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功取得會員資料
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    patch:
      tags:
        - User
      summary: 更新會員資料
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
                gender:
                  type: string
                  enum: [male, female, other]
                height:
                  type: number
                weight:
                  type: number
                targetWeight:
                  type: number
                activityLevel:
                  type: string
                  enum: [sedentary, light, moderate, active, very_active]
                dailyCalorieGoal:
                  type: number
                avatar:
                  type: string
                  format: uri
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'

    delete:
      tags:
        - User
      summary: 刪除帳號
      security:
        - BearerAuth: []
      responses:
        '204':
          description: 帳號已刪除
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: usr_1234567890
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: 王小明
        avatar:
          type: string
          format: uri
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        height:
          type: number
          description: 身高(公分)
        weight:
          type: number
          description: 體重(公斤)
        targetWeight:
          type: number
        activityLevel:
          type: string
          enum: [sedentary, light, moderate, active, very_active]
        dailyCalorieGoal:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: 請求參數有誤
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
...
    Unauthorized:
      description: 未授權或 Token 無效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...